pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["MrYazdan <mrrezayazdani@yahoo.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
python-decouple = "^3.8"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
